Add `delayStartup` argument to command line
This argument is useful in the situation that both TLSAttackerConnector
and the system under test are started at the same time. In this case,
the system under test is still initializing when TLSAttackerConnector
attempts to connect (which will fail and cause TLSAttackerConnector to
exit). Using a delay mitigates this issue.